algorithm/Algorithm/src/main/src/main/java/refractoring/Customer.java
package refractoring;

import java.util.*;

class Customer {
    private final String _name;
    private final Vector _rentals = new Vector();

    public Customer(String name) {
        _name = name;
    }

    public void addRental(Rental arg) {
        _rentals.addElement(arg);
    }

    public String getName() {
        return _name;
    }

    public String statement() {
        Enumeration rentals = _rentals.elements();
        String result = "Rental Record for " + getName() + "\n";
        while (rentals.hasMoreElements()) {
            Rental each = (Rental) rentals.nextElement();
            double thisAmount = each.get_movie().get_price().getCharge(each.getDaysRented());
            result += "\t" + each.getMovie().getTitle() + "\t" + thisAmount + "\n";
        }
        result += "Amount owed is " + getTotalCharge() + "\n";
        result += "You earned " + getTotalFrequentRenterPoints() + " frequent renter points";
        return result;
    }

    private double getTotalCharge() {
        double result = 0;
        Enumeration rentals = _rentals.elements();
        while (rentals.hasMoreElements()) {
            Rental each = (Rental) rentals.nextElement();
            result += each.get_movie().get_price().getCharge(each.getDaysRented());
        }
        return result;
    }

    private int getTotalFrequentRenterPoints() {
        int result = 0;
        Enumeration rentals = _rentals.elements();
        while (rentals.hasMoreElements()) {
            Rental each = (Rental) rentals.nextElement();
            result += each.get_movie().get_price().getFrequentRenterPoints(each.getDaysRented(), each.get_movie());
        }
        return result;
    }
}
